saluki_io::deser::framing

Trait FramerExt

pub trait FramerExt {
    // Required method
    fn framed<'a, F>(
        &'a mut self,
        framer: &'a mut F,
        is_eof: bool,
    ) -> Framed<'a, F, Self> 
       where Self: ReadIoBuffer + Sized,
             F: Framer;
}

Extension trait for ergonomically working with framers and buffers.

Required Methods

fn framed<'a, F>( &'a mut self, framer: &'a mut F, is_eof: bool, ) -> Framed<'a, F, Self>
where Self: ReadIoBuffer + Sized, F: Framer,

Creates a new Framed iterator over the buffer, using the given framer.

Returns an iterator that extracts frames from the given buffer, consuming the bytes from the buffer as frames are yielded.