saluki_io::net

Function build_datadog_agent_ipc_https_connector

pub async fn build_datadog_agent_ipc_https_connector<P: AsRef<Path>>(
    cert_path: P,
) -> Result<HttpsConnector<HttpConnector>, GenericError>

Builds an HTTPS connector for connecting to the Datadog Agent’s IPC endpoint.

This connector is configured to load a bundled TLS certificate/private key file, generated by the Datadog Agent, and use it for both server verification and client authentication. The given certificate path is expected to be a PEM-encoded file containing both the certificate and private key. The private key is only used for client authentication.