saluki_io::net::server::multiplex_service

Struct MultiplexService

pub struct MultiplexService<A, B> { /* private fields */ }

A Service that multiplexes requests between two underlying REST-ful and gRPC services.

In some scenarios, it can be useful to expose gRPC services on the same existing REST-ful HTTP endpoints as gRPC defaults to using HTTP/2, and doing so allows the use of a single exposed endpoint/port, leading to reduced configuration and complexity.

This service takes two services โ one for REST-ful requests and one for gRPC requests โ and multiplexes incoming requests between the two by inspecting the request headers, specifically the Content-Type header. If the header starts with application/grpc, the request is assumed to be a gRPC request and is forwarded to the gRPC service. Otherwise, the request is assumed to be a REST-ful request and is forwarded to the REST-ful service.

Implementations

impl<A, B> MultiplexService<A, B>

pub fn new(rest: A, grpc: B) -> Self

Creates a new MultiplexService from the given REST-ful and gRPC services.