saluki_io::net::util

Module retry

Structs

DiskUsageRetrieverImpl

ExponentialBackoff

An exponential backoff strategy.

NoopRetryPolicy

A no-op retry policy that never retries requests.

PushResult

Result of a push operation.

RetryQueue

A queue for storing requests to be retried.

RollingExponentialBackoffRetryPolicy

A rolling exponential backoff retry policy.

StandardHttpClassifier

A standard HTTP response classifier.

StandardHttpRetryLifecycle

A standard HTTP retry lifecycle that emits contextual information about HTTP requests and responses..

Traits

EventContainer

A container that holds events.

RetryClassifier

Determines whether or not a request should be retried.

Retryable

A value that can be retried.

Type Aliases

DefaultHttpRetryPolicy

A batteries-included retry policy suitable for HTTP-based clients.